+++ thelper/transforms/__init__.py
@@ -20,6 +20,7 @@ from thelper.transforms.operations import NoTransform  # noqa: F401
 from thelper.transforms.operations import RandomResizedCrop  # noqa: F401
 from thelper.transforms.operations import RandomShift  # noqa: F401
 from thelper.transforms.operations import Resize  # noqa: F401
+from thelper.transforms.operations import SelectChannels  # noqa: F401
 from thelper.transforms.operations import Tile  # noqa: F401
 from thelper.transforms.operations import ToColor  # noqa: F401
 from thelper.transforms.operations import ToGray  # noqa: F401